+++ ChangeLog.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 # Release 3.24 - 08/04/2024
 
+- FIX : display totalht ligne 'non compris' - *12/04/2024* - 3.24.3  
 - FIX : Title extrafields wasn't working - *12/04/2024* - 3.24.2  
 - FIX : doublon affichage label description lors de la création de facture depuis un objet (propal/commande ...)  - *03/04/2024* - 3.24.1  
 - NEW : Ajout des quantités par sous totaux sur l'interface - *27/03/2024* - 3.24.0